+++ src/unidaplan/DeletePTParameter.java
@@ -1,16 +1,13 @@
 package unidaplan;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
-
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.json.JSONException;
-import org.json.JSONObject;
+
 
 	public class DeletePTParameter extends HttpServlet {
 		private static final long serialVersionUID = 1L;
@@ -19,8 +16,13 @@ import org.json.JSONObject;
 	  public void doDelete(HttpServletRequest request, HttpServletResponse response)
 	      throws ServletException, IOException {		
 	    
-//		Authentificator authentificator = new Authentificator();
-//		int userID=authentificator.GetUserID(request,response);
+		Authentificator authentificator = new Authentificator();
+		int userID=authentificator.GetUserID(request,response);
+		
+		DBconnection dBconn = null;
+	    PreparedStatement pStmt = null;
+
+		
 		request.setCharacterEncoding("utf-8");
 	    int id=0;
 	  	  try  {
@@ -32,16 +34,20 @@ import org.json.JSONObject;
 	    String status="ok";
 
 	    try {
-	    // Delete the parameter
-	 	DBconnection DBconn=new DBconnection();
-	    DBconn.startDB();	   
-	    PreparedStatement pstmt = null;
-			pstmt= DBconn.conn.prepareStatement( 			
-					"DELETE FROM p_parameters WHERE id=? \n");
-		   	pstmt.setInt(1, id);
-		   	pstmt.executeUpdate();
-			pstmt.close();
-			DBconn.closeDB();
+		    // Delete the parameter
+		 	dBconn=new DBconnection();
+		    dBconn.startDB();	   
+		    
+			if (Unidatoolkit.userHasAdminRights(userID, dBconn)){
+				pStmt= dBconn.conn.prepareStatement( 			
+						"DELETE FROM p_parameters WHERE id=? \n");
+			   	pStmt.setInt(1, id);
+			   	pStmt.executeUpdate();
+				pStmt.close();
+			} else {
+				response.setStatus(401);
+			}
+		
 		} catch (SQLException e) {
 			System.err.println("DeletePTParameter: Problems with SQL query");
 			status="SQL Error; DeletePTParameter";
@@ -49,18 +55,10 @@ import org.json.JSONObject;
 			System.err.println("DeletePTParameter: Strange Problems");
 			status="Error DeletePTParameter";
 		}	
+		dBconn.closeDB();
+
 		
 	    // tell client that everything is fine
-		response.setContentType("application/json");
-	    response.setCharacterEncoding("utf-8");
-	    PrintWriter out = response.getWriter();
-	    try {
-	        JSONObject answer = new JSONObject();
-			answer.put("status", status);
-			answer.put("id", id);
-			out.println(answer.toString());
-		} catch (JSONException e) {
-			System.err.println("DeletePTParameter: Problems creating JSON answer");
-		}    
+		Unidatoolkit.sendStandardAnswer(status, response);   
 	}
 }
