+++ src/unidaplan/DeleteProcessFromExperiment.java
@@ -19,8 +19,11 @@ import org.json.JSONObject;
 	  public void doDelete(HttpServletRequest request, HttpServletResponse response)
 	      throws ServletException, IOException {		
 	    
-//		Authentificator authentificator = new Authentificator();
-//		int userID=authentificator.GetUserID(request,response);
+		Authentificator	authentificator = new Authentificator();
+		int userID = authentificator.GetUserID(request,response);
+	   	String privilege = "n";
+	   	PreparedStatement pStmt	= null;
+
 		request.setCharacterEncoding("utf-8");
 	    int id=0;
 	  	  	try{
@@ -32,23 +35,38 @@ import org.json.JSONObject;
 	    String status="ok";
 
 	    try {
-		    // Delete the user to the database	    
-		 	DBconnection DBconn=new DBconnection();
-		    DBconn.startDB();	   
-		    // decrease positions behind the sample to delete
-		    PreparedStatement pstmt = DBconn.conn.prepareStatement(    		
-	    		"UPDATE exp_plan_processes SET position = position - 1 "
-		    	+"WHERE expp_id=(SELECT expp_id FROM exp_plan_processes WHERE id=?) " 
-		    	+"AND position>(SELECT position FROM exp_plan_processes WHERE id=?)");
-	    	pstmt.setInt(1,id);
-	    	pstmt.setInt(2,id);
-	    	pstmt.executeUpdate();
-			pstmt= DBconn.conn.prepareStatement( 			
-					"DELETE FROM exp_plan_processes WHERE id=?");
-		   	pstmt.setInt(1, id);
-		   	pstmt.executeUpdate();
-			pstmt.close();
-			DBconn.closeDB();
+		    // Connect to database	    
+		 	DBconnection dBconn=new DBconnection();
+		    dBconn.startDB();	   
+		    
+		    // Check privileges
+		    pStmt = dBconn.conn.prepareStatement( 	
+					"SELECT getExperimentRights(vuserid:=?,vexperimentid:=?)");
+			pStmt.setInt(1,userID);
+			pStmt.setInt(2,id);
+			privilege = dBconn.getSingleStringValue(pStmt);
+			pStmt.close();
+			
+			if (privilege.equals("w")){
+		    
+		    
+			    // decrease positions behind the sample to delete
+			    pStmt = dBconn.conn.prepareStatement(    		
+		    		"UPDATE exp_plan_processes SET position = position - 1 "
+			    	+"WHERE expp_id=(SELECT expp_id FROM exp_plan_processes WHERE id=?) " 
+			    	+"AND position>(SELECT position FROM exp_plan_processes WHERE id=?)");
+		    	pStmt.setInt(1,id);
+		    	pStmt.setInt(2,id);
+		    	pStmt.executeUpdate();
+				pStmt= dBconn.conn.prepareStatement( 			
+						"DELETE FROM exp_plan_processes WHERE id=?");
+			   	pStmt.setInt(1, id);
+			   	pStmt.executeUpdate();
+				pStmt.close();
+			} else {
+				response.setStatus(401);
+			}
+			dBconn.closeDB();
 		} catch (SQLException e) {
 			System.err.println("DeleteProcessFromExperiment: Problems with SQL query");
 			status="SQL Error; DeleteProcessFromExperiment";
