+++ src/unidaplan/DeleteSTParameterGrp.java
@@ -2,6 +2,7 @@ package unidaplan;
 import java.io.IOException;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -15,8 +16,9 @@ import javax.servlet.http.HttpServletResponse;
 	  public void doDelete(HttpServletRequest request, HttpServletResponse response)
 	      throws ServletException, IOException {		
 	    
-//		Authentificator authentificator = new Authentificator();
-//		int userID=authentificator.GetUserID(request,response);
+		Authentificator authentificator = new Authentificator();
+		int userID=authentificator.GetUserID(request,response);
+		
 		String status="ok";
 		request.setCharacterEncoding("utf-8");
 	    int id=0;
@@ -29,17 +31,25 @@ import javax.servlet.http.HttpServletResponse;
 	   		System.err.print("DeleteSTParameterGrp: no parametergroup ID given!");
 	   	  }
 
-	    try {
-	    // Delete the user to the database	    
-	 	DBconnection DBconn=new DBconnection();
-	    DBconn.startDB();	   
-	    PreparedStatement pstmt = null;
-			pstmt= DBconn.conn.prepareStatement( 			
-					"DELETE FROM ot_parametergrps WHERE id=? ");
-		   	pstmt.setInt(1, id);
-		   	pstmt.executeUpdate();
-			pstmt.close();
-			DBconn.closeDB();
+	    try {   
+	    
+		// connect to database
+	 	DBconnection dBconn=new DBconnection();
+	    dBconn.startDB();
+	    
+	    if (Unidatoolkit.userHasAdminRights(userID, dBconn)){
+		    // Delete the user to the database	    
+
+		    PreparedStatement pStmt = null;
+				pStmt= dBconn.conn.prepareStatement( 			
+						"DELETE FROM ot_parametergrps WHERE id=? ");
+			   	pStmt.setInt(1, id);
+			   	pStmt.executeUpdate();
+				pStmt.close();
+				dBconn.closeDB();
+		    }else{
+		    	response.setStatus(401);
+		    }
 		} catch (SQLException e) {
 			System.err.println("DeleteSTParameterGrp: Problems with SQL query");
 			status="SQL Error; DeleteSTParameterGrp";
